refactor: change name and path for scripts
